requirements.txt
gunicorn==20.1.0
Flask~=1.1.2
PyGithub~=1.55
bs4~=0.0.1
beautifulsoup4~=4.9.3